agent/requirements.txt
requests>=2.31
google-generativeai>=0.8
python-dateutil>=2.8
